Add files via upload
Objective: Multi-class classification between forests, buildings, and sea using CNN In the dataset, there are three files: the forest, the building file, and the sea file. Each file consists of forest, building, and sea images. Then we are going to create a CNN architecture and predict on test data images and on new images as well.